Moved dummy data array to external file
